fix(ledger)!: widen committee expiry epochs to uint64
UpdateCommitteeGovAction.CredEpochs held absolute expiry epochs as uint.
The CDDL encodes that map value as `epoch = uint .size 8`, so on a 32-bit
build a conforming update_committee action could not carry an expiry at or
above 2^32. CommitteeMember.ExpiryEpoch, which stores the enacted form of
the same quantity, is already uint64; epoch_interval parameters such as
CommitteeTermLimit stay uint64 and MaxEpoch stays uint.

CredEpochs and the NewUpdateCommitteeGovAction parameter become
map[*Credential]uint64, dropping the widening conversions in
validateCommitteeTerms and ToPlutusData. This also lets the committee term
overflow case assert a full uint64 expiry on GOARCH=386.

+++ ledger/common/gov_constructors_test.go
@@ -149,7 +149,7 @@ func TestNewUpdateCommitteeGovAction(t *testing.T) {
 		Credential: NewBlake2b224([]byte("test")),
 	}
 	creds := []Credential{cred}
-	credEpochs := map[*Credential]uint{&cred: 42}
+	credEpochs := map[*Credential]uint64{&cred: 42}
 
 	// Negative: a zero-value Rat has a nil inner value and panics when CBOR
 	// encoded, so the constructor must reject it up front.
@@ -180,7 +180,7 @@ func TestNewUpdateCommitteeGovAction(t *testing.T) {
 	_, err = NewUpdateCommitteeGovAction(
 		nil,
 		creds,
-		map[*Credential]uint{nil: 1},
+		map[*Credential]uint64{nil: 1},
 		cbor.Rat{Rat: big.NewRat(2, 3)},
 	)
 	require.Error(t, err)
@@ -189,7 +189,7 @@ func TestNewUpdateCommitteeGovAction(t *testing.T) {
 func TestUpdateCommitteeGovActionUnmarshalCBORRejectsNilCredEpochKey(t *testing.T) {
 	encoded, err := cbor.Encode(UpdateCommitteeGovAction{
 		Type:       uint(GovActionTypeUpdateCommittee),
-		CredEpochs: map[*Credential]uint{nil: 42},
+		CredEpochs: map[*Credential]uint64{nil: 42},
 		Quorum:     cbor.Rat{Rat: big.NewRat(2, 3)},
 	})
 	require.NoError(t, err)
